datapipeline/requirements.txt
kafka-python==2.0.2
requests==2.31.0
pyyaml==6.0.1
Pillow